Implement `withFeaturedItem`
