+++ assets/js/blocks/featured-items/image-editor.js
@@ -3,7 +3,6 @@
 /**
  * External dependencies
  */
-import { useEffect, useState } from 'react';
 import {
 	__experimentalImageEditingProvider as ImageEditingProvider,
 	__experimentalImageEditor as GutenbergImageEditor,
@@ -55,9 +54,9 @@ export const ImageEditor = ( {
 };
 
 export const withImageEditor = ( Component ) => ( props ) => {
-	const [ isEditingImage, setIsEditingImage ] = useState( false );
+	const [ isEditingImage, setIsEditingImage ] = props.useEditingImage;
 
-	const { attributes, isSelected, name, setAttributes } = props;
+	const { attributes, backgroundImageSize, name, setAttributes } = props;
 	const { mediaId, mediaSrc } = attributes;
 	const item =
 		name === BLOCK_NAMES.featuredProduct ? props.product : props.category;
@@ -69,15 +68,11 @@ export const withImageEditor = ( Component ) => ( props ) => {
 		blockName: name,
 	} );
 
-	useEffect( () => {
-		setIsEditingImage( false );
-	}, [ isSelected ] );
-
 	if ( isEditingImage ) {
 		return (
 			<ImageEditor
 				backgroundImageId={ backgroundImageId }
-				backgroundImageSize={ { width: 500, height: 500 } }
+				backgroundImageSize={ backgroundImageSize }
 				backgroundImageSrc={ backgroundImageSrc }
 				isEditingImage={ isEditingImage }
 				setAttributes={ setAttributes }
@@ -86,10 +81,5 @@ export const withImageEditor = ( Component ) => ( props ) => {
 		);
 	}
 
-	return (
-		<Component
-			{ ...props }
-			useEditingImage={ [ isEditingImage, setIsEditingImage ] }
-		/>
-	);
+	return <Component { ...props } />;
 };
