+++ assets/js/blocks/featured-items/with-editing-image.js
@@ -1,0 +1,20 @@
+/**
+ * External dependencies
+ */
+import { useEffect, useState } from 'react';
+
+export const withEditingImage = ( Component ) => ( props ) => {
+	const [ isEditingImage, setIsEditingImage ] = useState( false );
+	const { isSelected } = props;
+
+	useEffect( () => {
+		setIsEditingImage( false );
+	}, [ isSelected ] );
+
+	return (
+		<Component
+			{ ...props }
+			useEditingImage={ [ isEditingImage, setIsEditingImage ] }
+		/>
+	);
+};
